post service impl done

+++ src/main/java/com/example/demo/models/Post.java
@@ -10,7 +10,7 @@ public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(updatable = false,nullable = false)
-    private Integer id;
+    private Long id;
     private String name;
 
     @Column(columnDefinition = "text")
@@ -27,7 +27,7 @@ public class Post {
     public Post() {
     }
 
-    public Post(Integer id, String name, String caption, String location, int likes, Date postedDate, Integer userImageId, List<Comment> commentList) {
+    public Post(Long id, String name, String caption, String location, int likes, Date postedDate, Integer userImageId, List<Comment> commentList) {
         this.id = id;
         this.name = name;
         this.caption = caption;
@@ -38,11 +38,11 @@ public class Post {
         this.commentList = commentList;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
